+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-None yet
+### Fixed
+
+- Fixed the `Subject::getParts()` method.
 
 ## [2.0.1] - 2019-01-09